refactor: ProjectService Command와 Query 서비스 분리
